Move utils.dispatchRequest() to FetchHttpClient
