+++ lib/policies/exponentialRetryPolicy.ts
@@ -118,7 +118,7 @@ export class ExponentialRetryPolicy extends BaseRequestPolicy {
     if (!utils.objectIsNull(response) && self.shouldRetry(response.status, retryData)) {
       try {
         await utils.delay(retryData.retryInterval);
-        const res: HttpOperationResponse = await utils.dispatchRequest(operationResponse.request);
+        const res: HttpOperationResponse = await this._nextPolicy.sendRequest(operationResponse.request);
         return self.retry(res, retryData, err);
       } catch (err) {
         return self.retry(operationResponse, retryData, err);
